+++ pkg/crypto/asymmetric/map_pubkeys_test.go
@@ -7,25 +7,21 @@ import (
 func TestMapPubKeys(t *testing.T) {
 	t.Parallel()
 
-	list := NewMapPubKeys()
-
 	pubKeys := []IPubKey{
 		NewPrivKey().GetPubKey(),
 		NewPrivKey().GetPubKey(),
 		NewPrivKey().GetPubKey(),
 	}
 
-	for _, pk := range pubKeys {
-		list.SetPubKey(pk)
-	}
-
-	if ok := list.InPubKeys(pubKeys[1]); !ok {
+	mapping := NewMapPubKeys(pubKeys...)
+	pkHash := pubKeys[1].GetHasher().ToBytes()
+	if pk := mapping.GetPubKey(pkHash); pk == nil {
 		t.Error("get invalid pub key")
 		return
 	}
 
-	list.DelPubKey(pubKeys[1])
-	if ok := list.InPubKeys(pubKeys[1]); ok {
+	mapping.DelPubKey(pubKeys[1])
+	if pk := mapping.GetPubKey(pkHash); pk != nil {
 		t.Error("get success deleted pub key")
 		return
 	}