+++ pkg/network/anonymity/examples/ping-pong/main.go
@@ -31,7 +31,7 @@ var (
 		numBytes = encoding.Uint64ToBytes(num + 1)
 		_ = n.SendPayload(
 			ctx,
-			pubKey.GetKEMPubKey(),
+			pubKey,
 			payload.NewPayload64(uint64(nodeRouter), numBytes[:]),
 		)
 		return nil, nil
@@ -73,12 +73,12 @@ func runServiceNode() anonymity.INode {
 	return node
 }
 
-func exchangeKeys(node1, node2 anonymity.INode) (asymmetric.IKEMPubKey, asymmetric.IKEMPubKey) {
+func exchangeKeys(node1, node2 anonymity.INode) (asymmetric.IPubKey, asymmetric.IPubKey) {
 	pubKey1 := node1.GetMessageQueue().GetClient().GetPrivKey().GetPubKey()
 	pubKey2 := node2.GetMessageQueue().GetClient().GetPrivKey().GetPubKey()
 
 	node1.GetMapPubKeys().SetPubKey(pubKey2)
 	node2.GetMapPubKeys().SetPubKey(pubKey1)
 
-	return pubKey1.GetKEMPubKey(), pubKey2.GetKEMPubKey()
+	return pubKey1, pubKey2
 }