+++ pkg/network/anonymity/logger/log_builder.go
@@ -15,7 +15,7 @@ type sLogBuilder struct {
 	fHash    []byte
 	fProof   uint64
 	fSize    uint64
-	fPubKey  asymmetric.IDSAPubKey
+	fPubKey  asymmetric.IPubKey
 	fConn    conn.IConn
 
 	fGetter ILogGetter
@@ -48,7 +48,7 @@ func (p *sLogBuilder) WithProof(pProof uint64) ILogBuilder {
 	return p
 }
 
-func (p *sLogBuilder) WithPubKey(pPubKey asymmetric.IDSAPubKey) ILogBuilder {
+func (p *sLogBuilder) WithPubKey(pPubKey asymmetric.IPubKey) ILogBuilder {
 	p.fPubKey = pPubKey
 	return p
 }